refactor(Utility): Move utility functions to their own folder, separate from scripts
Also removes the function for converting semantic selectors to attribute selectors which were incomplete (and thus confusing).
